\name{meanC}
\alias{meanC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate mean using C
}            
\description{
Calculate mean of a numerical vector C
}
\usage{
meanC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric values
}
}
\details{
Calculate the mean using C loop                                                           l
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
Input integer vector
}
\author{
Mahabubur Rahman
}
