\name{varC}
\alias{varC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate variance using C                      
}
\description{
Calculate mean of a numerical vector C
}
\usage{
varC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric values
}
}
\details{
Calculate the variance using C loop 
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
Input integer vector
}
\author{
Mahabubur Rahman
}
